feat(/): Make it possible to log every command (bump 0.4.0.0)
The user has to pass a `LogFn` to `runNixAction`, and can decide
whether to log something whenever a command is called.

This is helpful for enabling debug output within a custom monad.

It kinda forced us to move towards making `NixAction` a monad
transformer, but doing so feels like a good step in any case.

Also reordered the documentation for the types module to make it a bit
nicer to read.